requirements.txt
# FastAPI和服务相关
fastapi==0.104.1
uvicorn==0.24.0

# 数据验证和模型
pydantic==2.5.0

# 数据处理
pandas==2.1.3

# LangChain相关
langchain==0.1.0
langchain-core==0.1.0
langchain-openai==0.0.2

# 工具库
python-dotenv==1.0.0
loguru==0.7.2
tiktoken==0.5.1

# 可选：如果需要使用本地模型
# torch==2.1.0
# transformers==4.35.0
